change: Remove logger
The logger package has a lot of issues.
The colors are not working in web and on iOS.
Emojis are not working on some platforms.
It introduced a not very important library in a SDk which is intended to
be as minimal as possible.
This simplifies the logging to a format which works everywhere.
The # Syntax makes it also look good as markdown.
It is not very far away from the default syntax of errors and stacktraces.
It still provides several log levels, the same API and a list of output events for the logviewer.